refactor(tool-repair): split standalone parse helpers
